pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anti-turing"
version = "0.1.0"
description = "A terminal-based reverse Turing test game — blend in among AIs to survive"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.25.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]
keywords = ["game", "turing", "ai", "terminal", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
]

[project.scripts]
anti-turing = "anti_turing.main:main"

[tool.hatch.build.targets.wheel]
packages = ["anti_turing"]
